+++ src/app/shared/module/font-awesome.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from "@angular/core";
-import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
-import { library } from "@fortawesome/fontawesome-svg-core";
+import {
+  FontAwesomeModule,
+  FaIconLibrary
+} from "@fortawesome/angular-fontawesome";
 import {
   faAndroid,
   faApple,
@@ -14,13 +16,12 @@ import { faEnvelope, faLink } from "@fortawesome/free-solid-svg-icons";
   exports: [FontAwesomeModule]
 })
 export class FontAwesomeIconsModule {
-  constructor() {
-    library.add(faAndroid);
-    library.add(faApple);
-    library.add(faLinkedin);
-    library.add(faGithub);
+  constructor(library: FaIconLibrary) {
+    // Separated by library for ease of understanding
+    // Brand icons
+    library.addIcons(faAndroid, faApple, faLinkedin, faGithub);
 
-    library.add(faEnvelope);
-    library.add(faLink);
+    // Solid icons
+    library.addIcons(faEnvelope, faLink);
   }
 }